# backend/docs/openapi.apifox.cn.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Light OSS API",
    "version": "1.0.0",
    "description": "供 APIFOX 导入的 OpenAPI 3.0 JSON。文档中的路径与 HTTP 方法集合与后端实际注册的 Gin 路由保持一致。"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地后端默认地址"
    }
  ],
...
  "paths": {
...
    "/api/v1/buckets/{bucket}/objects/batch": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "post": {
        "tags": ["对象"],
        "operationId": "uploadObjectBatch",
        "summary": "使用 multipart/form-data 批量上传文件",
        "description": "`manifest` 是在 multipart 字段中编码的 JSON 数组。每个项目都声明 `file_field` 与 `relative_path`。其他文件分段必须使用 `manifest[].file_field` 中引用的精确字段名。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AllowOverwriteHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadBatchForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传的文件",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBatchEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "multipart 请求体、manifest、存储桶名称、前缀、可见性或最终对象键无效。错误代码可能包括 invalid_multipart_request、invalid_batch_manifest、batch_file_missing、invalid_bucket_name、invalid_prefix、invalid_visibility。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到存储桶。错误代码： bucket_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "存在同名对象且未允许覆盖。错误代码：object_exists。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "请求正文超过配置的上传大小。错误代码： payload_too_large。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "批量上传失败。错误代码可能包括 bucket_lookup_failed、object_lookup_failed、batch_file_buffer_failed、batch_file_open_failed、object_store_failed、object_metadata_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "BucketParam": {
        "name": "bucket",
        "in": "path",
        "required": true,
        "description": "存储桶名称。必须匹配 `^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$`。",
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$",
          "minLength": 3,
          "maxLength": 63
        }
      },
...
      "AllowOverwriteHeader": {
        "name": "X-Allow-Overwrite",
        "in": "header",
        "required": false,
        "description": "是否允许覆盖同对象键的已存在对象。省略时默认为 `false`。",
        "schema": {
          "type": "string",
          "enum": ["true", "false"]
        }
      },
...
    },
...
    "schemas": {
...
      "SuccessEnvelope": {
        "type": "object",
        "required": ["request_id", "data"],
        "properties": {
          "request_id": {
            "type": "string",
            "description": "中间件返回的请求关联ID。"
          },
          "data": {
            "type": "object",
            "description": "接口特定的响应数据。"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["request_id", "error"],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "Object": {
        "type": "object",
        "required": [
          "id",
          "bucket_name",
          "object_key",
          "original_filename",
          "size",
          "content_type",
          "etag",
          "visibility",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "bucket_name": {
            "type": "string"
          },
          "object_key": {
            "type": "string",
            "maxLength": 512
          },
          "original_filename": {
            "type": "string",
            "maxLength": 255
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "content_type": {
            "type": "string"
          },
          "etag": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": ["public", "private"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UploadBatchManifestItem": {
        "type": "object",
        "required": ["file_field", "relative_path"],
        "properties": {
          "file_field": {
            "type": "string",
            "description": "多部分文件字段名称。"
          },
          "relative_path": {
            "type": "string",
            "description": "提供的前缀下的相对对象路径。不得以“/”开头。"
          }
        }
      },
      "UploadBatchForm": {
        "type": "object",
        "required": ["manifest"],
        "properties": {
          "prefix": {
            "type": "string",
            "maxLength": 512,
            "description": "可选文件夹前缀。非空值必须以“/”结尾。"
          },
          "visibility": {
            "type": "string",
            "enum": ["public", "private"],
            "description": "可见性应用于每个上传的对象。省略时默认为private。"
          },
          "manifest": {
            "type": "array",
            "maxItems": 2000,
            "items": {
              "$ref": "#/components/schemas/UploadBatchManifestItem"
            },
            "description": "在 multipart 字段中编码的 UploadBatchManifestItem 对象 JSON 数组。",
            "example": [
              {
                "file_field": "file_0",
                "relative_path": "index.html"
              },
              {
                "file_field": "file_1",
                "relative_path": "assets/app.js"
              }
            ]
          }
        },
        "additionalProperties": {
          "type": "string",
          "format": "binary",
          "description": "由 `manifest[].file_field` 引用的其他 multipart 文件分段。"
        }
      },
      "UploadBatchResult": {
        "type": "object",
        "required": ["uploaded_count", "items"],
        "properties": {
          "uploaded_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Object"
            }
          }
        }
      },
...
      "UploadBatchEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/UploadBatchResult"
              }
            }
          }
        ]
      }
    }
  }
}